libdxfrw: and more compiler warnings

+++ libraries/libdxfrw/src/drw_entities.h
@@ -242,7 +242,7 @@ public:
 public:
     double staangle;               /*!< x coordinate, code 50 */
     double endangle;               /*!< x coordinate, code 51 */
-    double isccw;                  /*!< is counter clockwise arc?, only used in hatch, code 73 */
+    int isccw;                  /*!< is counter clockwise arc?, only used in hatch, code 73 */
 };
 
 //! Class to handle ellipse entity
@@ -265,7 +265,7 @@ public:
     double ratio;           /*!< ratio, code 40 */
     double staparam;        /*!< start parameter, code 41, 0.0 for full ellipse*/
     double endparam;        /*!< end parameter, code 42, 2*PI for full ellipse */
-    double isccw;           /*!< is counter clockwise arc?, only used in hatch, code 73 */
+    int isccw;           /*!< is counter clockwise arc?, only used in hatch, code 73 */
 };
 
 //! Class to handle trace entity
